Move Rgb struct into color.rs file
Instead of having the Rgb struct in the crate root, where it arguably
shoudn't be, it has been moved to the src/term/color.rs file.
